requirements.txt
torch==1.7.1
torchvision==0.5.0
numpy==1.18.3
gevent==20.9.0
Flask==1.1.2
requests==2.25.0
Werkzeug==1.0.1
Pillow==8.0.1
